requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.0
black==23.3.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.69
deta==1.1.0
fastapi==0.95.0
flake8==6.0.0
h11==0.14.0
idna==3.4
lxml==4.9.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
phjob-scraper==0.1.1
platformdirs==3.2.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
requests==2.28.2
requests-toolbelt==0.10.1
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
